+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.5.17:
+* Move to GWT 2.6.1.
 * Add toString to ChannelLink and ChannelDescriptor to ease debugging.
 * Avoid using Class.getSimpleName() in EntitySubscriptionManagerImpl to support GWT 2.5.1.
 * Correct the namespace in ReplicantConfig so that the property is read correctly.